+++ photutils/psf/tests/test_photometry.py
@@ -487,7 +487,8 @@ def test_psf_boundary():
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_aperture_radius_value_error():
     """
-    Test psf_photometry with discrete PRF model at the boundary of the data.
+    Test that a ValueError is raised for tabular PSF models when
+    aperture_radius is not defined.
     """
 
     prf = DiscretePRF(test_psf, subsampling=1)
@@ -496,11 +497,13 @@ def test_aperture_radius_value_error():
                                     bkg_estimator=None, psf_model=prf,
                                     fitshape=7)
 
-    intab = Table(data=[[1], [1]], names=['x_0', 'y_0'])
-    with pytest.raises(ValueError) as err:
-        basic_phot(image=image, init_guesses=intab)
+    with pytest.raises(ValueError):
+        basic_phot(image=image)
 
-    assert 'aperture_radius is None' in str(err.value)
+    # with initial guesses, but without a "flux_0" column
+    intab = Table(data=[[1], [1]], names=['x_0', 'y_0'])
+    with pytest.raises(ValueError):
+        basic_phot(image=image, init_guesses=intab)
 
 
 @pytest.mark.skipif('not HAS_SCIPY')